Fix: Refactor Firebase imports for platform specificity
- Introduced a barrel file system for Firebase initialization to resolve web bundling errors with native Firebase SDKs.
- Renamed `lib/firebase.web.ts` to `lib/firebase.ts` to act as the default (web) Firebase config.
- Modified `lib/firebase.ts` to initialize and export both Firestore and Auth instances using the Firebase Web SDK.
- Updated `context/AuthContext.tsx` and `app/edit-recipe/[id].tsx` to import Firebase instances from the new barrel file ` '@/lib/firebase'`.
- Verified that `hooks/useRecipes.ts` already used the correct barrel import.
- Verified `lib/firebase.native.ts` correctly exports instances using `@react-native-firebase/*` for native platforms.

This change aims to ensure the web build uses web-compatible Firebase SDKs, while native builds use the `@react-native-firebase` libraries.
