+++ lib/firebase.ts
@@ -1,6 +1,7 @@
-// Fil: lib/firebase.web.ts
+// Fil: lib/firebase.ts (Previously firebase.web.ts)
 import { initializeApp, getApp, getApps, FirebaseOptions } from 'firebase/app';
 import { getFirestore, Firestore } from 'firebase/firestore';
+import { getAuth, Auth } from 'firebase/auth'; // Import Firebase Auth for web
 
 const firebaseConfig: FirebaseOptions = {
     apiKey: process.env.EXPO_PUBLIC_API_KEY,
@@ -30,12 +31,31 @@ function initializeDb(): Firestore | null {
 
 // Vi kaller funksjonen én gang og eksporterer resultatet.
 const db = initializeDb();
+let authInstance: Auth | null = null;
+
+try {
+    if (firebaseConfig.apiKey && firebaseConfig.projectId) {
+        const app = !getApps().length ? initializeApp(firebaseConfig) : getApp();
+        authInstance = getAuth(app);
+    } else {
+        // Error already logged by initializeDb for missing config
+    }
+} catch (error) {
+    console.error("CRITICAL ERROR initializing Firebase Auth for web:", error);
+}
+
 
 // Gir en tydelig loggmelding om suksess eller fiasko.
 if (db) {
-    console.log("SUCCESS: Firebase for web initialized and ready.");
+    console.log("SUCCESS: Firestore for web initialized and ready.");
 } else {
-    console.error("FAILURE: Database object ('db') could not be initialized.");
+    console.error("FAILURE: Firestore object ('db') for web could not be initialized.");
 }
 
-export { db };
+if (authInstance) {
+    console.log("SUCCESS: Firebase Auth for web initialized and ready.");
+} else {
+    console.error("FAILURE: Firebase Auth object ('authInstance') for web could not be initialized.");
+}
+
+export { db, authInstance };
