+++ package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@expo-google-fonts/inter": "^0.4.1",
     "@react-native-async-storage/async-storage": "1.23.1",
-    "@react-native-firebase/app": "^22.2.1",
     "@react-native-firebase/auth": "^22.2.1",
     "expo": "~51.0.14",
     "expo-constants": "~16.0.2",
